refactor(config): split agents.rs into modules
Split 561-line src/config/agents.rs into 5 modules under src/config/agents/

Refs #221
